Revert "refactor: extract RedactionLockedTooltip component"
This reverts commit 3ca8721a49554e80542f7a75439d69232df2a051.
